+++ Player.h
@@ -12,7 +12,7 @@ private:
     float playerSpeed;
 
     float maxFireRate;
-    float fireRateTimer;
+    double fireRateTimer;
 
     sf::RectangleShape boundingRectangle;
 
@@ -24,7 +24,7 @@ public:
     ~Player();
     void Initialize();
     void Load();
-    void Update(float deltaTime,Skeleton& skeleton,sf::Vector2f& mousePosition);
+    void Update(double deltaTime,Skeleton& skeleton,sf::Vector2f& mousePosition);
     void Draw(sf::RenderWindow& window);
 };
